+++ be/src/main/java/io/nzbee/resources/product/PriceFacetMapper.java
@@ -15,7 +15,7 @@ public class PriceFacetMapper implements IFacetMapper<Double> {
 		ef.setDesc(price.toString());
 		ef.setHierarchical(false);
 		ef.setValue(price.toString());
-		ef.setCount(new Long(0));
+		ef.setCount(new Long(0).intValue());
 		return ef;
 	}
